stock-sentiment-analysis/stock_generator/requirements.txt
boto3
pandas
kafka-python